refactor(desktop): split runner/cadence into google-sync-runner.ts
Task 13 follow-up. sync-service.ts crossed the 800-line max-lines
threshold after M4b's push-runtime wiring landed. Moving the runner
lifecycle (startGoogleCalendarSyncRunner, stopGoogleCalendarSyncRunner),
the poll cadence (reEvaluatePollCadence, getCurrentPollIntervalMs,
PUSH_BACKOFF_INTERVAL_MS), and the interval state into
google-sync-runner.ts keeps the module focused on pull/push/writeback
logic.

sync-service.ts re-exports every moved symbol so existing callers
(index.ts, calendar-handlers, session-teardown, device-registration,
the cadence vitest suite) keep their original import paths with no
churn. No behavior change — same 722 desktop tests and 386 sync-server
tests still green.
